# resiresi-backend/openapi.yaml
openapi: "3.1.0"
info:
  title: ResiResi Restaurant Reservation API
  version: "1.0.0"
  description: |
    Backend reservation API for the ResiResi platform. Used as the APIblaze proxy reference origin.
    Authenticate with either `Authorization: Bearer <key>` or `x-api-key: <key>`.
servers:
  - url: https://backend.resiresi.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Restaurant not found.
            request_id:
              type: string
              format: uuid
            details:
              type: object
              description: Zod flatten() output on validation errors
...
    ApiKey:
      type: object
      required: [id, name, prefix, role, created_at]
      properties:
        id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        prefix:
          type: string
          example: rsrsi_live_A7K2XQ
        role:
          type: string
          enum: [platform, owner, manager, host, diner_app]
        created_by:
          type: string
          format: uuid
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ApiKeyCreated:
      allOf:
        - $ref: '#/components/schemas/ApiKey'
        - type: object
          required: [plaintext]
          properties:
            plaintext:
              type: string
              description: Full API key string — shown once only
              example: "rsrsi_live_A7K2XQ.Zk9s..."
...
    ApiKeyList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'
paths:
...
  /restaurants/{restaurantId}/api-keys:
    parameters:
      - name: restaurantId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: listApiKeys
      summary: List API keys for restaurant (owner only)
      tags: [API Keys]
      responses:
        "200":
          description: API key list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createApiKey
      summary: Create API key for restaurant (owner only) — plaintext returned once
      tags: [API Keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name:
                  type: string
                role:
                  type: string
                  enum: [owner, manager, host, diner_app]
                expires_at:
                  type: string
                  format: date-time
      responses:
        "201":
          description: API key created with plaintext
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}/api-keys/{keyId}:
    parameters:
      - name: restaurantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      operationId: revokeApiKey
      summary: Revoke API key (owner only)
      tags: [API Keys]
      responses:
        "204":
          description: Revoked
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /platform/api-keys:
    get:
      operationId: listPlatformApiKeys
      summary: List all API keys (platform only)
      tags: [Platform]
      responses:
        "200":
          description: All API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createPlatformApiKey
      summary: Create platform API key (platform only)
      tags: [Platform]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                expires_at:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Platform key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /platform/api-keys/{keyId}:
    parameters:
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      operationId: revokePlatformApiKey
      summary: Revoke any API key (platform only)
      tags: [Platform]
      responses:
        "204":
          description: Revoked
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'